worked on k-means

+++ src/ca/ece/ubc/cpen221/mp5/Review.java
@@ -21,7 +21,7 @@ public class Review {
     
     final private String type;
     final private Map<String, Long> votes = new HashMap<String, Long>();
-    private double rating;
+    private long rating;
     
     /**
      * Create a review object.
